+++ src/main/java/io/github/ricoapon/readableregex/GroupBuilder.java
@@ -5,39 +5,50 @@ package io.github.ricoapon.readableregex;
  */
 public interface GroupBuilder {
     /**
-     * Starts a new capturing group. This is the same as {@code (}. This must be finalized with calling {@link #endGroup()}.
+     * Starts a new capturing group. This is the same as {@code (}. You must call {@link #endGroup()} somewhere after this method.
      * @return This builder.
      */
     ReadableRegex startGroup();
 
     /**
-     * Starts a new capturing group for a given name. This is the same as {@code (<name>}. This must be finalized with
-     * calling {@link #endGroup()}.
+     * Starts a new capturing group for a given name. This is the same as {@code (<name>}. You must call {@link #endGroup()}
+     * somewhere after this method.
      * @param groupName The name of the group.
      * @return This builder.
      */
     ReadableRegex startGroup(String groupName);
 
     /**
-     * Starts a non-capturing group for positive lookbehind. This is the same as {@code (?<=}.
+     * Starts a non-capturing group. This is the same as {@code (?:}. You must call {@link #endGroup()} somewhere after
+     * this method.
+     * @return This builder.
+     */
+    ReadableRegex startUnnamedGroup();
+
+    /**
+     * Starts a non-capturing group for positive lookbehind. This is the same as {@code (?<=}. You must call {@link #endGroup()}
+     * somewhere after this method.
      * @return This builder.
      */
     ReadableRegex startPositiveLookbehind();
 
     /**
-     * Starts a non-capturing group for negative lookbehind. This is the same as {@code (?<!}.
+     * Starts a non-capturing group for negative lookbehind. This is the same as {@code (?<!}. You must call {@link #endGroup()}
+     * somewhere after this method.
      * @return This builder.
      */
     ReadableRegex startNegativeLookbehind();
 
     /**
-     * Starts a non-capturing group for positive lookahead. This is the same as {@code (?=}.
+     * Starts a non-capturing group for positive lookahead. This is the same as {@code (?=}. You must call {@link #endGroup()}
+     * somewhere after this method.
      * @return This builder.
      */
     ReadableRegex startPositiveLookahead();
 
     /**
-     * Starts a non-capturing group for negative lookahead. This is the same as {@code (?!}.
+     * Starts a non-capturing group for negative lookahead. This is the same as {@code (?!}. You must call {@link #endGroup()}
+     * somewhere after this method.
      * @return This builder.
      */
     ReadableRegex startNegativeLookahead();