+++ src/main/java/io/github/ricoapon/readableregex/ReadableRegexPattern.java
@@ -8,13 +8,20 @@ import java.util.regex.Pattern;
  * Wrapper of {@link Pattern} for additional useful methods.
  */
 public interface ReadableRegexPattern {
-
     /**
      * Matches the regular expression to the text. See {@link Pattern#matcher(CharSequence)} for more information.
      * @param text The text to be matched.
      * @return {@link Matcher}
      */
     Matcher matches(String text);
+
+    /**
+     * @param text The text to be matched.
+     * @return {@code true} if the pattern matches the full text, else {@code false}.
+     */
+    default boolean matchesTextExactly(String text) {
+        return matches(text).matches();
+    }
 
     /**
      * @return All the {@link PatternFlag}s that are enabled on this pattern.